Move AppHost tests to separate file
